I don't really know what I'm doing
I more or less did some reorganization and wrote a bunch of data
structures for no real reason. I want to mess with being able to
directly utilize the (almost)isomorphism between instructions and the
bindings they define. Maybe non-constant bindings are actually a pointer
to the instruction that defined them.